requirements.txt
streamlit
httpx_oauth
google-generativeai
pyserial
